Changing class to userprefs to make it more clear and inject it where needed.
